pyproject.toml
[project]
name = "hancom-pdf"
version = "0.1.0"
description = "PDF Text Extraction & Chunking API using opendataloader-pdf"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.1",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "opendataloader-pdf[hybrid]>=2.2.0",
    "tiktoken>=0.8.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "aiosqlite>=0.22.1",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
